+++ theme/template/inc/custom-header.php
@@ -56,17 +56,13 @@ if (!function_exists('_s_header_style')):
     // If we get this far, we have custom styles. Let's do this.
     ?>
 		<style type="text/css">
-		<?php // Has the text been hidden?
-  if (!display_header_text()): ?>
+		<?php if (!display_header_text()): ?>
 			.site-title,
 			.site-description {
 				position: absolute;
 				clip: rect(1px, 1px, 1px, 1px);
 			}
-		<?php
-    // If the user has set a custom color for the text use that.
-    // If the user has set a custom color for the text use that.
-    ?>else: ?>
+		<?php else: ?>
 			.site-title a,
 			.site-description {
 				color: #<?php echo esc_attr($header_text_color); ?>;